+++ common/cpp/react/renderer/components/TrueSheetSpec/TrueSheetViewShadowNode.h
@@ -8,6 +8,8 @@
 
 namespace facebook::react {
 
+class EventDispatcher;
+
 JSI_EXPORT extern const char TrueSheetViewComponentName[];
 
 /*
@@ -22,6 +24,8 @@ class JSI_EXPORT TrueSheetViewShadowNode final
   using ConcreteViewShadowNode::ConcreteViewShadowNode;
 
  public:
+  using StateData = ConcreteViewShadowNode::ConcreteStateData;
+
   static ShadowNodeTraits BaseTraits() {
     auto traits = ConcreteViewShadowNode::BaseTraits();
     traits.set(ShadowNodeTraits::Trait::RootNodeKind);
@@ -29,6 +33,13 @@ class JSI_EXPORT TrueSheetViewShadowNode final
   }
 
   void adjustLayoutWithState();
+
+#if !defined(ANDROID)
+  void setEventDispatcher(std::weak_ptr<const EventDispatcher> dispatcher);
+
+ private:
+  StateData &getStateDataMutable();
+#endif
 };
 
 } // namespace facebook::react